Simplify `Insert` actions Implement `Move` actions Move PyMethodsAnalyzer to BugFinderInspection
